project2-rag/requirements.txt
streamlit
chromadb
anthropic
python-dotenv
pdfplumber
python-docx
openpyxl
